Feelings Box
Refatorei um código da criação da feelings box
