+++ Mini02/Mini02/Presentation/Components/StatisticBoxView.swift
@@ -75,48 +75,7 @@ class StatisticsBoxView: UIView {
         // Adiciona a UIView interna ao stack view
         stackView.addArrangedSubview(customView)
     }
-    
-    func addImageAndLabel(_ imageName: String, labelText: String) {
-        // Cria uma UIView interna
-        let customView = UIView()
-        customView.backgroundColor = .clear // Torna o background da UIView interna invisível
-        customView.layer.cornerRadius = 8   // Arredonda as bordas da UIView interna
-        customView.clipsToBounds = true
         
-        // Cria uma UIImageView para a imagem
-        let imageView = UIImageView()
-        imageView.image = UIImage(named: imageName)
-        imageView.contentMode = .scaleAspectFit
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        
-        // Cria um rótulo para o texto
-        let label = UILabel()
-        label.text = labelText
-        label.textColor = textColorForCurrentMode() // Define a cor do texto com base no modo atual
-        label.textAlignment = .left // Alinhe o texto à esquerda
-        label.translatesAutoresizingMaskIntoConstraints = false
-        
-        // Adiciona a imagem e o rótulo à UIView interna
-        customView.addSubview(imageView)
-        customView.addSubview(label)
-        
-        // Configura as constraints para a imagem e o rótulo
-        NSLayoutConstraint.activate([
-            imageView.leadingAnchor.constraint(equalTo: customView.leadingAnchor, constant: 10), // Margem à esquerda
-            imageView.topAnchor.constraint(equalTo: customView.topAnchor, constant: 10), // Margem superior
-            imageView.widthAnchor.constraint(equalToConstant: 30), // Largura da imagem
-            imageView.heightAnchor.constraint(equalToConstant: 30), // Altura da imagem
-            
-            label.leadingAnchor.constraint(equalTo: imageView.trailingAnchor, constant: 10), // Espaço entre a imagem e o rótulo
-            label.trailingAnchor.constraint(equalTo: customView.trailingAnchor, constant: -10), // Margem à direita
-            label.topAnchor.constraint(equalTo: customView.topAnchor, constant: 10), // Margem superior
-            label.bottomAnchor.constraint(equalTo: customView.bottomAnchor, constant: -10), // Margem inferior
-        ])
-        
-        // Adiciona a UIView interna ao stack view
-        stackView.addArrangedSubview(customView)
-    }
-    
     private func textColorForCurrentMode() -> UIColor {
         if traitCollection.userInterfaceStyle == .dark {
             return .white // Modo escuro, cor do texto branca
